Abstract support actions and improve tests
