+++ src/actions/actionTypes.js
@@ -1,15 +1,16 @@
 export const REQUEST_PETITIONS = 'REQUEST_PETITIONS';
 export const RECEIVE_PETITIONS = 'RECEIVE_PETITIONS';
 
-export const CLEAR_PETITION = 'CLEAR_PETITION';
-
 export const REQUEST_PETITION = 'REQUEST_PETITION';
 export const RECEIVE_PETITION = 'RECEIVE_PETITION';
 
-export const SUBMIT_PETITION = 'SUBMIT_PETITION';
+export const SUBMITTING_PETITION = 'SUBMITTING_PETITION';
 export const CREATED_PETITION = 'CREATED_PETITION';
-export const UPDATED_PETITION = 'UPDATED_PETITION';
 export const PUBLISHED_PETITION = 'PUBLISHED_PETITION';
+export const UPDATED_PETITION = 'UPDATED_PETITION';
+export const CLEAR_PETITION = 'CLEAR_PETITION';
+
+export const SUBMITTING_SUPPORT = 'SUBMITTING_SUPPORT';
 export const SUPPORTED_PETITION = 'SUPPORTED_PETITION';
 
 export const REQUEST_WHOAMI = 'REQUEST_WHOAMI';
