+++ src/reducers/petition.js
@@ -4,7 +4,8 @@ import {
   RECEIVE_PETITION,
   CREATED_PETITION,
   UPDATED_PETITION,
-  SUBMIT_PETITION,
+  SUBMITTING_PETITION,
+  SUBMITTING_SUPPORT,
   PUBLISHED_PETITION,
   SUPPORTED_PETITION
 } from 'actions/actionTypes';
@@ -33,21 +34,28 @@ export default function petition (state = initialState, action) {
           saved: true
         }
       );
-    case SUPPORTED_PETITION:
-      return Object.assign({},
-        state, action.petition,
-        { isLoading: false }
-      );
-    case SUBMIT_PETITION:
+    case SUBMITTING_PETITION:
       return Object.assign({},
         state, action.petition,
         { isLoading: true }
+      );
+    case SUBMITTING_SUPPORT:
+      return Object.assign({},
+        state,
+        { isSupporting: true }
       );
     case PUBLISHED_PETITION:
       return Object.assign({},
         state, action.petition, {
           published: true,
           isLoading: false
+        }
+      );
+    case SUPPORTED_PETITION:
+      return Object.assign({},
+        state, action.petition, {
+          hasSupported: true,
+          isSupporting: false
         }
       );
     case CLEAR_PETITION:
